Credify_files/requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.3.6
certifi==2025.7.14
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.36.0
Django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-celery-beat==2.7.0
drf-yasg==1.21.7
inflection==0.5.1
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
redis==5.0.3
sentry-sdk==1.40.6
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
django-environ==0.11.2
django-redis==5.4.0
gunicorn==21.2.0
requests>=2.31.0


